+++ examples/async_std.rs
@@ -1,6 +1,0 @@
-use async_main::async_main;
-
-#[async_main(async_std)]
-async fn main(_executor: ()) {
-    println!("Hello, world!");
-}