+++ examples/futures.rs
@@ -1,7 +1,0 @@
-use async_main::async_main;
-use futures::executor::LocalSpawner;
-
-#[async_main(futures)]
-async fn main(_executor: LocalSpawner) {
-    println!("Hello, world!");
-}