+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://github.com/AldaronLau/semver).
 
+## [0.2.0] - 2023-01-17
+### Added
+ - *`web`* feature 
+ - `Spawn` trait
+ - `LocalSpawner` struct
+
+### Changed
+ - Broke proc macro into separate crate
+ - Generate same code for all async runtimes
+ - Replace attribute parameters with features
+
+### Removed
+ - smolscale feature
+
 ## [0.1.0] - 2023-01-01
 ### Added
  - Tokio support